+++ internal/app/io_helpers.go
@@ -1,4 +1,4 @@
-package main
+package app
 
 import (
 	"bufio"
