+++ internal/app/ndc.go
@@ -1,4 +1,4 @@
-package main
+package app
 
 type ndcItem struct {
 	Code  string
