+++ internal/app/service.go
@@ -1,4 +1,4 @@
-package main
+package app
 
 import (
 	"context"
